+++ HISTORY.md
@@ -4,6 +4,9 @@ Change History
 @NEXT@
 ------
 
+- merged git@github.com:ekashida/yui-pathogen-decoder.git (default decoder)
+- relax yui version validation from 3.x.x to x.x.x
+
 0.3.4
 ------